# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "API Reserva de turnos",
        "description": "Esta es la documentacion de nsdnasdkasnd.",
        "version": "1.0.0"
    },
    "paths": {
...
        "/rest/turnos/{id}": {
            "get": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Obtener un turno por su ID",
                "description": "Retorna los detalles de un turno específico.",
                "operationId": "6d029c84080727215a4eb2bf0ed06a7d",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID del turno a obtener.",
                        "required": true,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Detalles del turno obtenidos exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "id": {
                                            "type": "integer",
                                            "example": "1"
                                        },
                                        "fecha_turno": {
                                            "type": "string",
                                            "format": "date",
                                            "example": "18/05/2023"
                                        },
                                        "hora_turno": {
                                            "type": "string",
                                            "format": "time",
                                            "example": "18:00:00"
                                        },
                                        "id_cancha": {
                                            "type": "integer",
                                            "example": 1
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Turno no encontrado.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "error": {
                                            "type": "string"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
        "/rest/turnos/buscarPorCancha/{idCancha}": {
            "get": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Buscar turnos por cancha",
                "description": "Retorna los turnos disponibles para una cancha específica.",
                "operationId": "97dd56812685c3db9901efdbf3b5fdf4",
                "parameters": [
                    {
                        "name": "idCancha",
                        "in": "path",
                        "description": "ID de la cancha para buscar los turnos.",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Lista de turnos para la cancha obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "fecha_turno": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "18/05/2023"
                                            },
                                            "hora_turno": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "18:00:00"
                                            },
                                            "id_cancha": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    }
                }
            }
        },
        "/api/turnos/searchByDate": {
            "post": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Buscar turnos por fecha",
                "description": "Retorna los turnos disponibles para una fecha específica.",
                "operationId": "d64dd3f7812396200e85957bb227b423",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "18/05/2023"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de turnos para la fecha especificada obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "fecha_turno": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "18/05/2023"
                                            },
                                            "hora_turno": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "18:00:00"
                                            },
                                            "id_cancha": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación. La fecha proporcionada no es válida."
                    }
                }
            }
        },
        "/api/turnos/searchByDateAndCategory": {
            "post": {
                "tags": [
                    "Turnos"
                ],
                "summary": "Buscar turnos por fecha y categoría",
                "description": "Retorna los turnos disponibles para una fecha específica y una categoría específica de cancha.",
                "operationId": "602a7e7f09d3ed1e27959157e5b0ad84",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "fecha": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "18/05/2023"
                                    },
                                    "id_categoria": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Lista de turnos para la fecha y categoría especificadas obtenida exitosamente.",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "properties": {
                                            "id": {
                                                "type": "integer",
                                                "example": "1"
                                            },
                                            "fecha_turno": {
                                                "type": "string",
                                                "format": "date",
                                                "example": "18/05/2023"
                                            },
                                            "hora_turno": {
                                                "type": "string",
                                                "format": "time",
                                                "example": "18:00:00"
                                            },
                                            "id_cancha": {
                                                "type": "integer",
                                                "example": 1
                                            }
                                        },
                                        "type": "object"
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "Error de validación. La fecha o la categoría proporcionada no son válidas."
                    }
                }
            }
        }
    }
}